Fixed MT06173: Software loaded outside of software list (e.g. -cart PATH_TO_ROM) fails to save SRAM.

+++ src/emu/diimage.cpp
@@ -161,16 +161,16 @@ image_error_t device_image_interface::set_image_filename(const char *filename)
 	util::zippath_parent(m_working_directory, filename);
 	m_basename.assign(m_image_name);
 
-	size_t loc1 = m_image_name.find_last_of('\\');
-	size_t loc2 = m_image_name.find_last_of('/');
-	size_t loc3 = m_image_name.find_last_of(':');
-	size_t loc = MAX(loc1,MAX(loc2, loc3));
+	int loc1 = m_image_name.find_last_of('\\');
+	int loc2 = m_image_name.find_last_of('/');
+	int loc3 = m_image_name.find_last_of(':');
+	int loc = MAX(loc1, MAX(loc2, loc3));
 	if (loc != -1) {
 		if (loc == loc3)
 		{
 			// temp workaround for softlists now that m_image_name contains the part name too (e.g. list:gamename:cart)
 			m_basename = m_basename.substr(0, loc);
-			size_t tmploc = m_basename.find_last_of(':');
+			int tmploc = m_basename.find_last_of(':');
 			m_basename = m_basename.substr(tmploc + 1, loc - tmploc);
 		}
 		else
